requirements.txt
fastapi
uvicorn
python-docx
pdfplumber
python-multipart
google-generativeai
